// src/client-mobile-v1/api-client/models/AuthSignUpDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Izi Crawler API
 * server-api-v1
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AuthSignUpDto
 */
export interface AuthSignUpDto {
    /**
     * 
     * @type {string}
     * @memberof AuthSignUpDto
     */
    email: string;
    /**
     * 
     * @type {string}
     * @memberof AuthSignUpDto
     */
    password: string;
    /**
     * 
     * @type {string}
     * @memberof AuthSignUpDto
     */
    confirmPassword: string;
}

/**
 * Check if a given object implements the AuthSignUpDto interface.
 */
export function instanceOfAuthSignUpDto(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "email" in value;
    isInstance = isInstance && "password" in value;
    isInstance = isInstance && "confirmPassword" in value;

    return isInstance;
}

export function AuthSignUpDtoFromJSON(json: any): AuthSignUpDto {
    return AuthSignUpDtoFromJSONTyped(json, false);
}

export function AuthSignUpDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): AuthSignUpDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'email': json['email'],
        'password': json['password'],
        'confirmPassword': json['confirmPassword'],
    };
}

export function AuthSignUpDtoToJSON(value?: AuthSignUpDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'email': value.email,
        'password': value.password,
        'confirmPassword': value.confirmPassword,
    };
}